upload file in team member show, pass meeting as hidden
