+++ app/views/meetings/show.html.erb
@@ -1,24 +1,5 @@
 <p id="notice"><%= notice %></p>
-<% if !@testt.nil? %>
-	<% @testt.each do |t| %>
-		<%= t %>
-	<% end %>
-<% end %>
-<p>
-  <h2><%= @meeting.title %></h2>
-</p>
 
-<p>
-  <%= @meeting.description %>
-</p>
-
-<%= form_tag uploader_path, :method => 'get' do %>
-  <%= hidden_field_tag :id, params[:id] %>
-  <%= hidden_field_tag :team_member_id, params[:id] %>
-  <label for="file">Select a calendar</label><%= file_field_tag "file" %>
-
-  <%= submit_tag 'Upload'%>
-<% end %>
 
 
 
@@ -31,6 +12,8 @@ Schedule goes here
 <a href="javascript:ReverseDisplay('share')">
 Click to share this meeting with whoever you'd like:
 </a>
+
+<%= link_to 'Upload your schedule', new_team_member_path(:meeting_id => @meeting.id) %>
 
 
 <div id="share" style="display:none;">
