+++ app/views/team_members/show.html.erb
@@ -1,5 +1,9 @@
 <p id="notice"><%= notice %></p>
-
+<% if !@events.nil? %>
+	<% @events.each do |event| %>
+		<%= event %>
+	<% end %>
+<% end %>
 <p>
   <strong>Name:</strong>
   <%= @team_member.name %>
@@ -14,6 +18,13 @@
   <strong>Meeting:</strong>
   <%= @team_member.meeting_id %>
 </p>
+
+<%= form_tag uploader_path, :method => 'get' do %>
+  <%= hidden_field_tag :id, params[:id] %>
+  <label for="file">Select a calendar</label><%= file_field_tag "file" %>
+
+  <%= submit_tag 'Upload'%>
+<% end %>
 
 <%= link_to 'Edit', edit_team_member_path(@team_member) %> |
 <%= link_to 'Back', team_members_path %>
