+++ config/routes.rb
@@ -6,7 +6,7 @@ Rails.application.routes.draw do
     # Routes for Home
   get '/' => 'home#home', as: :home
 
-  get 'uploader' => 'meetings#uploader', as: :uploader
+  get 'uploader' => 'team_members#uploader', as: :uploader
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
